cpu-boost: Bind to system-background cpus and reduce RT prio
To prevent the case of a possible RT priority inversion with
more important RT tasks like mdss (RIP), reduce the priority to 2.
This will make sure cpu-boost can take priority over most work
items while not causing a priority inversion leading to thread
interruption.

Additionally we bind the cpu-boost worker thread to CPUs 0-2 to
decrease the likeliness of it interfering with the app thread, which
preferably runs on cpu3 or the big cores.
For that purpose, we do a small refactor of the kthread init process.

While at it, add logging for the case where sched_setscheduler fails,
as it can fail due to security checks.

Change-Id: I03be69f5382919c8e18a0b927384d799965a6435